updated home page and nav bar
fixed the 'about me' on the home page. removed reference to photos since
I didn't end up launching them.

also fixed the svgs for instagram and github, as the file names were
mismatched with their references and therefore they didn't render.

also added browsersync to webpack, and moved my images to the resources
directory, adding a .copyDirectory to webpack in order to move them to
the public directory during compilation