started new post frontend
aded new post component, and implemented live markdown parsing so that I can write a markdown post on the page and it will show me a preview.

still need to wire up the submission logic, including adding a password so that only I can submit posts.

also added a landing page when someone tries to navigate to a route that doesn’t exist.